+++ server/views.py
@@ -1,58 +1,12 @@
 import os
-from server import app, db
-from server.desciptor import Descriptor
-from flask import render_template, request, jsonify
-from server.models import Authentication
+from server import app
+from flask import render_template, request
 from server.authentication import authenticate, development_only
 from server.atl_requests.lib import AtLib
 
 site_root = os.path.realpath(os.path.dirname(__file__))
 static_path = os.path.join(os.path.join(site_root, "static"))
 template_path = os.path.join(os.path.join(site_root, "templates"))
-
-
-@app.route('/')
-def index():
-    return "Welcome"
-
-
-@app.route('/install/<descriptor>')
-def install(descriptor):
-    """ Return the atlassian json file, supports multiple descriptors"""
-    install_app = {}
-    if descriptor == 'pythia':
-        install_app = Descriptor(
-            key='pythia-addon',
-            name='Pythia',
-            description='Visualize the current state of the engineering system and identify any engineering constraints',
-            vendor_name='Ryan Schaffer',
-            vendor_url='https://www.example.com',
-            scopes=['read', 'admin']
-        )
-
-    return jsonify(install_app.descriptor)
-
-
-@app.route('/add-on-installed-callback', methods=['POST'])
-def installed():
-    install_data = request.get_json()
-    install_data.update({'installedBy': request.args.get('user_key')})
-    auth = Authentication(install_data)
-    db.session.add(auth)
-    db.session.commit()
-    return 'OK'
-
-
-@app.route('/add-on-enable', methods=['POST'])
-def enabled():
-    return 'OK'
-
-
-@app.route('/uninstall', methods=['POST'])
-def uninstalled():
-    Authentication.query.filter_by(clientKey=request.get_json().get('clientKey')).delete()
-    db.session.commit()
-    return 'OK'
 
 
 @app.route('/welcome')
